requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
flask-migrate==4.1.0
python-dotenv==1.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
SQLAlchemy==2.0.41
WTForms==3.2.1
alembic==1.16.1
Pillow
reportlab==4.0.7
arabic-reshaper==3.0.0
python-bidi==0.4.2